refactor: Move firestore and sync code to utils sub folders
