+++ app/utils/firestore/README.md
@@ -1,4 +1,4 @@
 # MIGRATION: Firestore API
 
 Kann zum finalen Release endgültig entfernt werden. Showstopper ist eine funktionierende
-Import/Export-Lösung.
+Import/Export-Lösung. Bis dahin könnten die Firestore-Daten als Backup dienen.
