+++ app/utils/firestore/championship-player.ts
@@ -1,4 +1,4 @@
-import { firestore, modelConverter } from '#.server/firestore';
+import { firestore, modelConverter } from './firestore.server';
 
 export type ChampionshipPlayer = {
   id: string;
