+++ app/utils/firestore/championship-round.ts
@@ -1,4 +1,4 @@
-import { firestore, modelConverter } from '#.server/firestore';
+import { firestore, modelConverter } from './firestore.server';
 
 type Round = {
   id: string;
